move check_conn_addr() utility and reuse it
