+++ .github/workflows/build_wheel.yml
@@ -17,7 +17,7 @@ jobs:
       CIBW_TEST_COMMAND_MACOS: LC_ALL='en_US.utf8' python -Wa {project}/psutil/tests/runner.py
       CIBW_TEST_EXTRAS: test
       # https://cibuildwheel.readthedocs.io/en/stable/options/
-      CIBW_SKIP: cp35-* pp27-win32*
+      CIBW_SKIP: cp35-* pp*-win32* pp37*
     steps:
     - uses: actions/checkout@v2
     - uses: actions/setup-python@v2
